requirements.txt
asgiref==3.8.0
Django==4.2.7
django-filter==24.1
django-jazzmin==2.6.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.0
PyJWT==2.8.0
sqlparse==0.4.4
